+++ demo/gradio/computer_use_demo/agent/anthropic_agent.py
