+++ demo/gradio/computer_use_demo/agent/llm_utils/oai.py
